+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning 2.0.0](https://semver.org/spec/
 - User profile and dashboard backend
 - Threaded comments backend
 - Notifications backend
+- Real-time notifications using Socket.io
 
 ### Fixed